platform/tenant-management/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pyyaml==6.0
kubernetes==23.6.0
pytest==7.0.1
requests==2.27.1
sqlalchemy==1.4.36